Added main, networkclient and unpdated programstructure
Alot has changed here. It is beleived to be a better structure
