+++ RovControl/RovControl/Gui/RESOURCES/MainWindow.ui
